DROTH-3397 moved getTwoDigitLaneCode to LaneNumber
